manual-instrumentation/requirements.txt
Django==5.2.3
gunicorn==23.0.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-instrumentation-django==0.55b1
opentelemetry-exporter-otlp-proto-http==1.34.1
newrelic-telemetry-sdk==0.4.3
newrelic~=9.13.0
sentry-sdk==2.19.2
python-json-logger==3.3.0
